refactor: extract SiteAdapter interface and separate Gemini-specific DOM logic
